临时提交

+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/Line.cpp
@@ -17,11 +17,11 @@ Line::~Line() {
 
 }
 
-Vector Line::pointAt(const float &t) const {
+Vector Line::pointAt(const double &t) const {
     return Vector::add(Vector::scalarMultiply(_point1,1-t),Vector::scalarMultiply(_point2,t));
 }
 
-float Line::distance(const Vector& Q)const{
+double Line::distance(const Vector& Q)const{
     return line2ray(*this).distance(Q);
 }
