+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/Line.h
@@ -15,8 +15,8 @@ public:
     Line(const Vector& p1,const Vector& p2);
     ~Line();
 public:
-    Vector pointAt(const float& t)const;
-    float distance(const Vector& Q)const;
+    Vector pointAt(const double& t)const;
+    double distance(const Vector& Q)const;
     const Vector& point1()const;
     const Vector& point2()const;
     bool pointLiesInLine(const Vector& Q)const;
