+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/LineDistance.h
@@ -17,10 +17,10 @@ public:
 
 
 
-    static float calc(const Line& l1, const Line& l2);
-    static float calc(const Ray& l1,const Ray& l2);
-    static float calc(const Line& l1, const Ray& l2);
-    static float calc(const Ray& l1,const Line& l2);
+    static double calc(const Line& l1, const Line& l2);
+    static double calc(const Ray& l1,const Ray& l2);
+    static double calc(const Line& l1, const Ray& l2);
+    static double calc(const Ray& l1,const Line& l2);
 
 
     static bool isParallel(const Vector&,const Vector&);
