+++ android/sdk/lolila_sdk/src/main/cpp/base/geometry/Ray.cpp
@@ -22,21 +22,21 @@ Ray::~Ray() {
 
 }
 
-Vector Ray::pointAt(const float &t) const {
+Vector Ray::pointAt(const double &t) const {
     return Vector::add(_endpoint,Vector::scalarMultiply(_direction,t));
 }
 
-float Ray::distance(const Vector &Q) const {
+double Ray::distance(const Vector &Q) const {
     const Vector& S = _endpoint;
     const Vector& V = _direction;
     const Vector Q_S = Vector::minus(Q,S);
     const Vector proj_Q_S_V = Vector::proj(Q_S,V);
     // 如果f1和和f2相同的话，直接相减会出现误差，而不是0，因此需要使用工具函数
-    const float f1 = powf(Q_S.length(),2.0);
+    const double f1 = pow(Q_S.length(),2.0);
     //LogQueue::push("f1:%g\n",f1);
-    const float f2 = powf(proj_Q_S_V.length(),2.0);
+    const double f2 = pow(proj_Q_S_V.length(),2.0);
     //LogQueue::push("f2:%g\n",f2);
-    return sqrtf( FloatUtils::minus(f1,f2));
+    return sqrt( FloatUtils::minus(f1,f2));
 }
 
 const Vector& Ray::endPoint()const{
@@ -55,5 +55,5 @@ bool Ray::pointLiesInRay(const Vector& Q)const{
     const Vector proj_Q_S_V = Vector::proj(Q_S,V);
     return FloatUtils::isEqual(Q_S.length(),proj_Q_S_V.length());
      */
-   return FloatUtils::isEqual(distance(Q),0);
+   return FloatUtils::isEqual(distance(Q),0.0);
 }
